perf(chat): lazy-load assistant-ui runtime off the initial bundle
The global chat provider wraps every page, so its assistant-ui runtime
(~200 KB) and the thread list were statically imported into the entry chunk
and modulepreloaded on every initial load - including for logged-out visitors
on public pages who never render chat. Extract GrueneratorChatRuntimeProvider
into its own module loaded via React.lazy (the unauthenticated branch never
references it), and lazy-load ChatThreadList in GlobalChatProvider. Logged-out
landing-page loads no longer fetch assistant-ui.
